move private nsis stuff out of the common distributed data dir.
these libs and headers files are only needed in Windows, and do not belong
to the common `data` directory, which is distributed in all platforms.

I suspect, but I'm not sure, that they are only needed during production of
the nsis installer, and I'm sure that moving them out of the way as I just
did broke the installer.  please correct the paths there, do not put these
files back in the `data` directory.

relates to #109